requirements.txt
django
djangorestframework
gunicorn
opencv-python==4.2.0.34
